* fixes BUG ID: 3033335 * testing revised * other minor changes

+++ src/main/com/siemens/ct/exi/io/BitInputStream.java
@@ -95,7 +95,7 @@ final public class BitInputStream {
 	 * @param n
 	 * @throws IOException
 	 */
-	public void skip(int n) throws IOException {
+	public void skip(long n) throws IOException {
 		if (capacity == 0) {
 			//	aligned
 			while(n != 0) {
