+++ src/main/com/siemens/ct/exi/io/channel/ByteDecoderChannel.java
@@ -49,7 +49,7 @@ public class ByteDecoderChannel extends AbstractDecoderChannel implements
 	public void align() throws IOException {
 	}
 	
-	public void skip(int n) throws IOException {
+	public void skip(long n) throws IOException {
 		while(n != 0) {
 			n -= is.skip(n);
 		}
